+++ src/test/kotlin/study/Language.kt
@@ -1,0 +1,3 @@
+package study
+
+data class Language(val language: String, val level: Int)
