+++ src/test/kotlin/study/Skill.kt
@@ -1,0 +1,9 @@
+package study
+
+interface Skill {
+    val description: String
+}
+
+data class SoftSkill(override val description: String): Skill
+
+data class HardSkill(override val description: String): Skill
